Rename authServerService to authApi
